fix: Hide first wallet in onboard modal if mobile pairing is enabled (#3654)
* fix: Hide first wallet in onboard modal if mobile pairing is enabled

* fix: Return null instead of Fragment
