--- src/components/AppLayout/Header/components/ProviderDetails/hidePairingModule.css
@@ -1,3 +1,0 @@
-.bn-onboard-modal-select-wallets li:first-of-type {
-    display: none;
-}
